@@ -2,10 +2,10 @@
   "name": "mapwork",
   "version": "1.0.0",
   "description": "mapwork-legacy",
-  "main": "index.js",
+  "type": "module",
   "scripts": {
-    "dev": "parcel index.html",
-    "build": "parcel build index.html --no-source-maps --public-url ./ --no-content-hash",
+    "start": "vite",
+    "build": "vite build",
     "test": "jest --coverage",
     "test-ci": "jest --ci --coverage --reporters=default --reporters=jest-junit",
     "test:once": "jest",
@@ -22,49 +22,33 @@
     "@babel/core": "^7.7.7",
     "@babel/polyfill": "^7.8.7",
     "@babel/preset-env": "^7.7.7",
+    "@sveltejs/vite-plugin-svelte": "^1.0.0-next.30",
+    "@testing-library/jest-dom": "^5.16.3",
+    "@testing-library/svelte": "3.1.0",
+    "@tsconfig/svelte": "^2.0.1",
+    "@types/jest": "^27.4.1",
+    "@types/jquery": "^3.5.14",
     "eslint": "^6.8.0",
     "eslint-plugin-svelte3": "^2.7.3",
-    "jest": "^24.9.0",
+    "jest": "^27.5.1",
     "jest-canvas-mock": "^2.2.0",
+    "jest-circus": "^26.2.2",
     "jest-junit": "^10.0.0",
     "jquery": "^3.4.1",
-    "parcel": "^1.12.4",
-    "parcel-bundler": "^1.12.4",
-    "parcel-plugin-svelte": "^4.0.6",
     "prettier": "^2.0.0",
     "prettier-plugin-svelte": "^0.7.0",
     "sass": "^1.23.7",
-    "svelte": "^3.20.1",
-    "svelte-preprocess": "^3.5.0"
+    "svelte": "^3.44.0",
+    "svelte-check": "^2.2.7",
+    "svelte-jester": "^1.0.5",
+    "svelte-mock": "^1.0.1",
+    "svelte-preprocess": "^4.9.8",
+    "ts-jest": "^27.1.4",
+    "tslib": "^2.3.1",
+    "typescript": "^4.5.4",
+    "vite": "^2.9.0"
   },
   "browserslist": [
     "last 1 chrome versions"
-  ],
-  "jest": {
-    "collectCoverage": true,
-    "coveragePathIgnorePatterns": [
-      "/node_modules/"
-    ],
-    "setupFiles": [
-      "jest-canvas-mock",
-      "./src/jest.setup.js"
-    ],
-    "coverageReporters": [
-      "text",
-      "json",
-      "cobertura"
-    ],
-    "reporters": [
-      "default",
-      "jest-junit"
-    ]
-  },
-  "jest-junit": {
-    "output": "output/coverage/junit/junit.xml",
-    "usePathForSuiteName": true
-  },
-  "dependencies": {
-    "core-js": "^3.6.1",
-    "regenerator-runtime": "^0.13.3"
-  }
+  ]
 }
