@@ -1,6 +1,6 @@
-import { writable } from 'svelte/store'
+import { writable } from 'svelte/store';
 
 export const NotificationConfig = writable({
   message: null,
   colour: null,
-})
+});
