@@ -3,4 +3,4 @@ export const WizardTypes = {
   WIZARD_CREATE_MAP: 0,
   WIZARD_SAVE_MAP: 1,
   WIZARD_DOWNLOAD_MAP: 2,
-}
+};
